refactor(ci): isolate runner-loss GitHub evidence
